app/requirements.txt
Flask==2.2.5
psycopg2-binary
pytest==7.4.0
Werkzeug==2.2.3
gunicorn==20.1.0
